+++ dataproduct_apps/collect.py
@@ -2,82 +2,11 @@ import datetime
 import logging
 import os
 
-from k8s.base import Model
-from k8s.fields import Field, ListField
-from k8s.models.common import ObjectMeta
 
+from dataproduct_apps.crd import Application, Topic
 from dataproduct_apps.model import App, TopicAccessApp, AppRef, appref_from_rule
 
 LOG = logging.getLogger(__name__)
-
-
-class TokenX(Model):
-    enabled = Field(bool, False)
-
-
-class Rules(Model):
-    application = Field(str)
-    namespace = Field(str)
-    cluster = Field(str)
-
-
-class External(Model):
-    host = Field(str)
-
-
-class Inbound(Model):
-    rules = ListField(Rules)
-
-
-class Outbound(Model):
-    external = ListField(External)
-    rules = ListField(Rules)
-
-
-class AccessPolicy(Model):
-    inbound = Field(Inbound)
-    outbound = Field(Outbound)
-
-
-class ApplicationSpec(Model):
-    image = Field(str)
-    ingresses = ListField(str)
-    tokenx = Field(TokenX)
-    access_policy = Field(AccessPolicy)
-
-
-class Application(Model):
-    class Meta:
-        list_url = "/apis/nais.io/v1alpha1/applications"
-        url_template = "/apis/nais.io/v1alpha1/namespaces/{namespace}/applications/{name}"
-
-    apiVersion = Field(str, "nais.io/v1alpha1")  # NOQA
-    kind = Field(str, "Application")
-
-    metadata = Field(ObjectMeta)
-    spec = Field(ApplicationSpec)
-
-
-class TopicAccess(Model):
-    access = Field(str)
-    application = Field(str)
-    team = Field(str)
-
-
-class TopicSpec(Model):
-    pool = Field(str)
-    acl = ListField(TopicAccess)
-
-
-class Topic(Model):
-    class Meta:
-        list_url = "http://localhost:8001/apis/kafka.nais.io/v1/topics"
-        url_template = "/apis/nais.io/v1/namespaces/{namespace}/topics/{name}"
-
-    apiVersion = Field(str, "kafka.nais.io/v1")  # NOQA
-    kind = Field(str, "Topic")
-    metadata = Field(ObjectMeta)
-    spec = Field(TopicSpec)
 
 
 def init_k8s_client():
@@ -124,9 +53,9 @@ def parse_apps(collection_time, cluster, apps, topics):
         outbound_apps = []
         outbound_hosts = []
         for rule in app.spec.access_policy.inbound.rules:
-            inbound_apps = inbound_apps + [appref_from_rule(cluster, metadata.namespace, rule).as_string()]
+            inbound_apps = inbound_apps + [str(appref_from_rule(cluster, metadata.namespace, rule))]
         for rule in app.spec.access_policy.outbound.rules:
-            outbound_apps.append(appref_from_rule(cluster, metadata.namespace, rule).as_string())
+            outbound_apps.append(str(appref_from_rule(cluster, metadata.namespace, rule)))
         for host in app.spec.access_policy.outbound.external:
             outbound_hosts.append(host.host)
         app = App(
